ポケモンリストとタイプリストの取得機能を追加し、関連するAPIエンドポイントを実装
