+++ src/app/(root)/components/PokemonList/index.tsx
@@ -1,0 +1,23 @@
+import { Pokemon } from "@/types/data";
+import { request } from "@/utils/request";
+import Link from "next/link";
+
+export async function PokemonList() {
+	const response = await request<Pokemon[]>(
+		"http://localhost:3000/api/pokemonList",
+	);
+
+	if (response.ok === false || response.data === undefined) {
+		return <p>データの取得に失敗しました</p>;
+	}
+
+	return (
+		<ul>
+			{response.data.map((pokemon) => (
+				<li key={pokemon.id}>
+					<Link href={`/pokemon/${pokemon.id}`}>{pokemon.name}</Link>
+				</li>
+			))}
+		</ul>
+	);
+}
