--- src/app/api/pokemonInfo/route.ts
@@ -1,6 +1,0 @@
-import { getPokemonInfo } from "@/data/pokemonInfo";
-
-export async function GET(request: Request) {
-	const pokemonInfo = await getPokemonInfo();
-	return Response.json(pokemonInfo);
-}
