+++ src/app/api/pokemonList/route.ts
@@ -1,0 +1,6 @@
+import { getPokemonList } from "@/data/pokemonList";
+
+export async function GET(request: Request) {
+	const pokemonList = await getPokemonList();
+	return Response.json(pokemonList);
+}
