+++ src/app/api/typeList/route.ts
@@ -1,0 +1,6 @@
+import { getTypeList } from "@/data/typeList";
+
+export async function GET(request: Request) {
+	const typeList = await getTypeList();
+	return Response.json(typeList);
+}
