+++ src/app/pokemon/[id]/page.tsx
@@ -1,13 +1,13 @@
 type Props = {
-	params: Promise<{ pokemonId: number }>;
+	params: Promise<{ id: number }>;
 };
 
 export default async function Page({ params }: Props) {
-	const { pokemonId } = await params;
+	const { id } = await params;
 	return (
 		<main>
 			<h1>ポケモン情報</h1>
-			<p>id: {pokemonId}</p>
+			<p>id: {id}</p>
 		</main>
 	);
 }
