+++ src/data/typeList.ts
@@ -1,0 +1,5 @@
+import { Type } from "@/types/data";
+
+export async function getTypeList(): Promise<Type[]> {
+	return ["くさ", "ほのお", "みず", "でんき", "ノーマル"];
+}
