+++ src/types/data.d.ts
@@ -1,5 +1,7 @@
+export type Type = "くさ" | "ほのお" | "みず" | "でんき" | "ノーマル";
+
 export type Pokemon = {
 	id: number;
 	name: string;
-	type: string;
+	type: Type;
 };
